Move arch-specific code to separate files instead of overusing the preprocessor
