+++ arch/aarch64/dispatcher_aarch64.S
@@ -4,7 +4,7 @@
 
   Copyright 2013-2016 Cosmin Gorgovan <cosmin at linux-geek dot org>
   Copyright 2015-2017 Guillermo Callaghan <guillermocallaghan at hotmail dot com>
-  Copyright 2017 The University of Manchester
+  Copyright 2017-2020 The University of Manchester
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -21,74 +21,6 @@
 .global start_of_dispatcher_s
 start_of_dispatcher_s:
 
-#ifdef __arm__
-.syntax unified
-
-.global th_to_arm
-.func th_to_arm
-.thumb_func
-th_to_arm:
-  bx pc
-.endfunc
-#endif // __arm__
-
-.global dispatcher_trampoline
-.func   dispatcher_trampoline
-
-#ifdef __arm__
-.code 32
-dispatcher_trampoline:
-  #R2 is available at this point
-  #TODO: INSTALL our own stack
-
-#A subroutine must preserve the contents of the registers r4-r8, r10, r11 and SP (and r9 in PCS variants that designate r9 as v6).
-  PUSH {r3 - r6, r9, r12, lr}
-  STR R0, [R3, #-4] // save the SPC
-  SUB R2, R3, #8    // set up the TPC pointer
-  VPUSH {d16-d31}
-  VPUSH {d0-d7}
-
-  MRS r5, CPSR
-  VMRS r6, FPSCR
-
-  LDR R3, disp_thread_data
-  LDR R9, dispatcher_addr
-
-  # provide 8-byte alignment of the SP
-  MOV R4, SP
-  BIC SP, #0x7
-  BLX R9
-  MOV SP, R4
-
-  MSR CPSR, r5
-  VMSR FPSCR, r6
-
-  VPOP {d0-d7}
-  VPOP {d16-d31}
-  POP {r3 - r6, r9, r12, lr}
-
-/*  SP ->  R0
-           R1
-           R2
-           R3
-       <0 or more words>
-           EMPTY
-           TPC
-           SPC
-    R3 ->
-  */
-
-  LDR R1, [SP, #12]
-  STR R1, [R3, #-12]
-  POP {R0, R1, R2}
-  SUB SP, R3, #12
-  POP {R3}
-  B checked_cc_return
-
-dispatcher_addr: .word dispatcher
-#endif
-
-#ifdef __aarch64__
 .global push_neon
 push_neon:
   STP  Q0,  Q1, [SP, #-512]!
@@ -154,9 +86,8 @@ pop_x4_x21:
   LDP X20, X21, [SP, #128]
   LDP  X4,  X5, [SP], #144
   RET
-#endif
 
-#ifdef __aarch64__
+.global dispatcher_trampoline
 dispatcher_trampoline:
   // PUSH all general purpose registers but X0, X1
   // X0 and X1 are pushed by the exit stub
@@ -189,61 +120,9 @@ dispatcher_trampoline:
   B checked_cc_return
 
 dispatcher_addr: .quad dispatcher
-#endif
-.endfunc
+
 
 .global trace_head_incr
-.func   trace_head_incr
-#ifdef __arm__
-.code 32
-  SUB PC, PC, #3
-.thumb_func
-trace_head_incr:
-  STR LR, [SP, #16]
-  NOP // MOVW R1, #counter_base & 0xFFFF    
-  NOP
-  NOP // MOVT R1, #counter_base >> 16
-  NOP
-  LDRB R2, [R1, R0]
-  SUBW  R2, R2, #1
-  STRB R2, [R1, R0]
-  CBZ  R2, create_trace_trampoline
-  POP {R0-R2, LR}
-  LDR PC, [SP], #8
-
-.align 2
-create_trace_trampoline:
-  BX PC
-  NOP
-.code 32
-  ADD R2, SP, #16
-  PUSH {R3 - R6, R9, R12}
-  VPUSH {D16-D31}
-  VPUSH {D0-D7}
-
-  MRS R5, CPSR
-  VMRS R6, FPSCR
-
-  MOV R1, R0
-  LDR R0, disp_thread_data
-  LDR R3, =create_trace
-
-  MOV R4, SP
-  BIC SP, #0x7
-  BLX R3
-  MOV SP, R4
-
-  MSR CPSR, R5
-  VMSR FPSCR, R6
-
-  VPOP {D0-D7}
-  VPOP {D16-D31}
-  POP {R3 - R6, R9, R12}
-
-  POP {R0-R2, LR}
-  B checked_cc_return
-#endif // __arm__
-#ifdef __aarch64__
 trace_head_incr:
   /*
    * X1 = Basic Block number
@@ -302,69 +181,9 @@ create_trace_trampoline:
 
   B checked_cc_return
 
-#endif // __aarch64__
-.endfunc
 
 .global syscall_wrapper
 .global syscall_wrapper_svc
-.func   syscall_wrapper
-#ifdef __arm__
-.code 32
-syscall_wrapper:
-  # R8 is the SPC of the following instruction
-  # R14 is the address where to return in the code cache
-  STR LR, [SP, #56] // save the TPC
-  STR R8, [SP, #60] // save the SPC
-
-  MOV R0, R7 // syscall id
-  MOV R1, SP // pointer to saved regs
-  MOV R2, R8 // SPC of the next instr.
-  LDR R3, disp_thread_data
-
-  LDR R4, syscall_handler_pre_addr
-  // provide 8-byte alignment of the SP
-  MOV R5, SP
-  BIC SP, #0x7
-  BLX R4
-  MOV SP, R5
-
-  // don't execute the syscall if pre handler returns 0
-  CMP R0, #0
-  BEQ s_w_r
-
-  // only allow overriding R0-R7
-  // the value of R8 must be preserved
-  LDM SP, {R0-R7}
-
-  // Balance the stack on sigreturn and rt_sigreturn, which don't return here
-  CMP R7, #119
-  ADDEQ SP, SP, #64
-  CMP R7, #173
-  ADDEQ SP, SP, #64
-
-  SVC 0
-syscall_wrapper_svc:
-  STM SP, {R0-R7}
-
-  MOV R0, R7
-  MOV R1, SP
-  MOV R2, R8
-  LDR R3, disp_thread_data
-
-  LDR R4, syscall_handler_post_addr
-  // provide 8-byte alignment of the SP
-  MOV R5, SP
-  BIC SP, #0x7
-  BLX R4
-  MOV SP, R5
-s_w_r: POP {R0-R12, R14}
-  B checked_cc_return
-
-syscall_handler_pre_addr: .word syscall_handler_pre
-syscall_handler_post_addr: .word syscall_handler_post
-#endif // __arm__
-
-#ifdef __aarch64__
 syscall_wrapper:
   STP X30, X29, [SP, #-16]!
   BL push_x4_x21
@@ -425,90 +244,14 @@ s_w_r:
 
 syscall_handler_pre_addr: .quad syscall_handler_pre
 syscall_handler_post_addr: .quad syscall_handler_post
-#endif // __aarch64__
-.endfunc
 
 .global disp_thread_data
-disp_thread_data:
-#ifdef __arm__
-  .word 0
-#endif
-#ifdef __aarch64__
-  .quad 0
-#endif
+disp_thread_data: .quad 0
 
 .global send_self_signal
 
 .global checked_cc_return
-.func   checked_cc_return
 checked_cc_return:
-#ifdef __arm__
-.code 32
-  SUB PC, PC, #3
-.thumb_func
-  PUSH {R0}
-  LDR R0, th_is_pending_ptr
-  LDR R0, [R0]
-  CBZ R0, gotocc
-  B deliver_signals_trampoline
-gotocc:
-  POP {R0}
-  LDR PC, [SP], #8
-deliver_signals_trampoline:
-  LDR R0, [SP, #8] // load SPC
-  SUB SP, SP, #12
-  PUSH {r1 - r2}
-  PUSH {r3 - r6, r9, r12, lr}
-  ADD R1, SP, #36
-  VPUSH {d16-d31}
-  VPUSH {d0-d7}
-
-  MRS r5, CPSR
-  VMRS r6, FPSCR
-
-  LDR R9, =deliver_signals
-
-  # provide 8-byte alignment of the SP
-  MOV R4, SP
-  BIC R2, r4, #0x7
-  MOV SP, R2
-  BLX R9
-  MOV SP, R4
-
-  MSR CPSR, r5
-  VMSR FPSCR, r6
-
-  VPOP {d0-d7}
-  VPOP {d16-d31}
-  POP {r3 - r6, r9, r12, lr}
-
-  /*  SP-> r1
-           r2
-           PID
-           TID
-           SIGNO
-           R0
-           TPC
-           SPC
-  */
-
-  CBZ R0, abort_self_signal
-
-  PUSH {R7}
-
-  ADD R2, SP, #12
-  LDM R2, {R0-R2}
-  MOV R7, #268
-  SVC 0
-send_self_signal:
-  POP {r7}
-abort_self_signal:
-  POP {r1-r2}
-  ADD SP, SP, #12
-  POP {r0}
-  LDR PC, [SP], #8
-
-#elif __aarch64__
   STR X2, [SP, #-16]!
   LDR X2, th_is_pending_ptr
   LDR W2, [X2]
@@ -569,15 +312,9 @@ abort_self_signal:
   LDR X2, [SP, #16]
   LDR X0, [SP], #32
   BR X0
-#endif
-.endfunc
 
 .global th_is_pending_ptr
-#ifdef __arm__
-th_is_pending_ptr: .word 0
-#elif __aarch64__
 th_is_pending_ptr: .quad 0
-#endif
 
 # place the literal pool before the end_of_dispatcher_s symbol
 .ltorg
