+++ arch/aarch64/scanner_a64.c
